устраняет недостатки из предыдущего коммита
