+++ js/main.js
@@ -3,4 +3,4 @@ createPhotoDescription();
 import './usersPhoto.js';
 import './bigPhoto.js';
 import './userForm.js';
-import './validationHashtagComment.js';
+import './validationForm.js';
